fix new circular import
