+++ packages/api-domain/src/domain/contexts/user/personal-user/index.ts
@@ -5,7 +5,7 @@ export type {
 } from './personal-user.entity.ts';
 export type { PersonalUserRepository } from './personal-user.repository.ts';
 export type { PersonalUserUnitOfWork } from './personal-user.uow.ts';
-export type { PersonalUserAccountProps } from './personal-user-account.ts';
-export type { PersonalUserProfileProps } from './personal-user-account-profile.ts';
-export type { PersonalUserAccountProfileLocationProps } from './personal-user-account-profile-location.ts';
-export type { PersonalUserAccountProfileBillingProps } from './personal-user-account-profile-billing.ts';
+export type { PersonalUserAccountProps } from './personal-user-account.entity.ts';
+export type { PersonalUserProfileProps } from './personal-user-account-profile.entity.ts';
+export type { PersonalUserAccountProfileLocationProps } from './personal-user-account-profile-location.entity.ts';
+export type { PersonalUserAccountProfileBillingProps } from './personal-user-account-profile-billing.entity.ts';
