+++ packages/api-domain/src/domain/contexts/user/personal-user/personal-user-account-profile-billing.entity.ts
@@ -1,0 +1,10 @@
+import type { DomainSeedwork } from '@cellix/domain-seedwork';
+
+export interface PersonalUserAccountProfileBillingProps
+	extends DomainSeedwork.ValueObjectProps {
+	subscriptionId?: string | undefined;
+	cybersourceCustomerId?: string | undefined;
+}
+
+export interface PersonalUserAccountProfileBillingEntityReference
+	extends Readonly<PersonalUserAccountProfileBillingProps> {}
