+++ packages/api-domain/src/domain/contexts/user/personal-user/personal-user-account-profile-location.entity.ts
@@ -1,0 +1,14 @@
+import type { DomainSeedwork } from '@cellix/domain-seedwork';
+
+export interface PersonalUserAccountProfileLocationProps
+	extends DomainSeedwork.ValueObjectProps {
+	address1: string;
+	address2?: string | undefined;
+	city: string;
+	state: string;
+	country: string;
+	zipCode: string;
+}
+
+export interface PersonalUserAccountProfileLocationEntityReference
+	extends Readonly<PersonalUserAccountProfileLocationProps> {}
