+++ packages/api-domain/src/domain/contexts/user/personal-user/personal-user-account.ts
@@ -3,22 +3,9 @@ import type { UserVisa } from '../user.visa.ts';
 import type { PersonalUserAggregateRoot } from './personal-user.ts';
 import { PersonalUserProfile } from './personal-user-account-profile.ts';
 import type {
-	PersonalUserProfileProps,
-	PersonalUserProfileEntityReference,
-} from './personal-user-account-profile.ts';
-export interface PersonalUserAccountProps
-	extends DomainSeedwork.ValueObjectProps {
-	accountType: string;
-	email: string;
-	username: string;
-
-	readonly profile: PersonalUserProfileProps;
-}
-
-export interface PersonalUserAccountEntityReference
-	extends Readonly<Omit<PersonalUserAccountProps, 'profile'>> {
-	readonly profile: PersonalUserProfileEntityReference;
-}
+	PersonalUserAccountEntityReference,
+	PersonalUserAccountProps,
+} from './personal-user-account.entity.ts';
 
 export class PersonalUserAccount
 	extends DomainSeedwork.ValueObject<PersonalUserAccountProps>
